updates from team review (#31)
* schema mod for multi-attester sync

* new favicon

* fix footer

* refactor EpochKeyInfo/Event components

* remove jsx body tag

* refactor epoch#/transition time

* clean up css

* remove hash from epochKey page
