+++ packages/frontend/public/index.html
@@ -11,7 +11,7 @@
       rel="stylesheet"
       href="https://fonts.googleapis.com/css2?family=Azeret+Mono:wght@200..700&display=swap"
     />
-    <title>explorer</title>
+    <title>UniRep Explorer</title>
     <script>
       window.__DEV_CONFIG__ = '__DEV_CONFIG__'
     </script>
