+++ packages/frontend/src/App.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { Routes, Route } from 'react-router-dom'
-import Header from './components/Header'
 import Home from './pages/Home'
 import AttesterPage from './pages/AttesterPage'
 import UserPage from './pages/UserPage'
@@ -11,13 +10,11 @@ import './app.css'
 export default function App() {
   return (
     <Routes>
-      <Route path="/" element={<Header />}>
-        <Route index element={<Home />} />
-        <Route path="attester/:id" element={<AttesterPage />} />
-        <Route path="user/:id" element={<UserPage />} />
-        <Route path="epochKey/:id" element={<EpochKeyPage />} />
-        <Route path="*" element={<NotFound />} />
-      </Route>
+      <Route index element={<Home />} />
+      <Route path="attester/:id" element={<AttesterPage />} />
+      <Route path="user/:id" element={<UserPage />} />
+      <Route path="epochKey/:id" element={<EpochKeyPage />} />
+      <Route path="*" element={<NotFound />} />
     </Routes>
   )
 }
