+++ packages/frontend/src/app.css
@@ -6,47 +6,47 @@ body {
 }
 
 h1 {
-  font-size: 3em;
+  font-size: 3rem;
   font-weight: 700;
-  margin: 0.1em 0;
+  margin: 0.1rem 0;
 }
 
 h2 {
-  font-size: 2em;
+  font-size: 2rem;
   font-weight: 700;
-  margin: 0.2em 0;
+  margin: 0.2rem 0;
 }
 
 h3 {
-  font-size: 1.5em;
+  font-size: 1.5rem;
   font-weight: 700;
-  margin: 0.5em 0;
+  margin: 0.5rem 0;
   color: #494e53;
 }
 
 h4 {
-  font-size: 0.9em;
+  font-size: 0.9rem;
   font-weight: 600;
-  margin: 0.4em 0;
+  margin: 0.4rem 0;
   color: #669294;
 }
 
 h5 {
-  font-size: 0.9em;
+  font-size: 0.9rem;
   font-weight: 200;
-  margin: 0.4em 0;
+  margin: 0.4rem 0;
 }
 
 h6 {
-  font-size: 0.9em;
+  font-size: 0.9rem;
   font-weight: 600;
-  margin: 0.4em 0;
+  margin: 0.4rem 0;
 }
 
 button {
   background-color: white;
-  border-radius: 24px;
-  padding: 6px 30px;
+  border-radius: 1.5rem;
+  padding: 0.3rem 1.8rem;
   cursor: pointer;
 }
 
@@ -65,14 +65,24 @@ li {
   align-items: center;
 }
 
+.content {
+  min-height: 100vh;
+  display: grid;
+  grid-template-rows: auto 1fr auto;
+}
+
 .container {
   display: flex;
-  margin: 50px 50px;
+  margin: 3.5rem;
+}
+
+.footer {
+  padding: 4.5rem 3.5rem;
 }
 
 .left-container {
   width: 30%;
-  margin-right: 5%;
+  margin-right: 4rem;
 }
 
 .right-container {
@@ -82,21 +92,12 @@ li {
 .info-grid {
   display: grid;
   grid-template-columns: 50% 50%;
-  gap: 12px;
-  margin-bottom: 8px;
-}
-
-.graph-container {
-  width: 100%;
-  height: auto;
-  border: 1px solid black;
-  border-radius: 8px;
-  background-color: #a0d8db;
-  margin-bottom: 8%;
+  gap: 0.75rem;
+  margin-bottom: 3.5rem;
 }
 
 .events-header {
-  padding: 20px 10px;
+  padding: 0.6rem;
 }
 
 .selected {
@@ -107,27 +108,4 @@ li {
 .unselected {
   color: #9ea3a7;
   cursor: pointer;
-}
-
-.scroll {
-  width: auto;
-  max-height: 400px;
-  overflow-x: hidden;
-  margin-bottom: 5%;
-}
-
-.green {
-  color: #89bd5f;
-}
-
-.red {
-  color: #e17f75;
-}
-
-.dot {
-  height: 15px;
-  width: 15px;
-  background-color: #89bd5f;
-  border-radius: 50%;
-  display: inline-block;
 }
