+++ packages/frontend/src/components/EpochKeyEvent.jsx
@@ -1,6 +1,5 @@
-import React, { useContext, useEffect } from 'react'
+import React from 'react'
 import { observer } from 'mobx-react-lite'
-import state from '../contexts/state'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import './eventCard.css'
@@ -8,27 +7,13 @@ import './eventCard.css'
 dayjs.extend(relativeTime)
 
 export default observer(({ attestation }) => {
-  const { attesterId, epoch, fieldIndex, change, timestamp } = attestation
-  const { attester } = useContext(state)
-  useEffect(() => {
-    const loadData = async () => {
-      await attester.loadEpochsByAttester(attesterId)
-    }
-    loadData()
-  }, [])
-  const attesterEpoch = attester.epochByNumber(attesterId, epoch)
+  const { fieldIndex, change, timestamp } = attestation
 
   return (
     <div className="event-card">
-      {attesterEpoch ? (
-        attesterEpoch.sealed ? (
-          <p>sealed</p>
-        ) : (
-          <p>not sealed</p>
-        )
-      ) : (
-        'Loading...'
-      )}
+      {/* <p>
+        {_id.slice(0, 5)}...{_id.slice(-5)}
+      </p> */}
       <p>{fieldIndex}</p>
       <p style={{ minWidth: '20px' }}>{'0x' + BigInt(change).toString(16)}</p>
       <p style={{ minWidth: '100px', fontSize: '12px' }}>
