+++ packages/frontend/src/components/eventCard.css
@@ -4,7 +4,7 @@
   align-items: center;
   width: auto;
   border: 1px solid black;
-  border-radius: 8px;
-  margin-bottom: 1em;
-  padding: 0 1em;
+  border-radius: 0.5rem;
+  margin-bottom: 1rem;
+  padding: 0 1rem;
 }
