--- packages/frontend/src/components/footer.css
@@ -1,5 +1,0 @@
-.footer {
-  /* background-image: url("../footer_bg.svg"); */
-  /* background-color: rgb(174, 170, 170); */
-  padding: 4.5em 3em;
-}
