+++ packages/frontend/src/components/header.css
@@ -11,32 +11,32 @@ a:visited {
 .header {
   display: flex;
   justify-content: space-between;
-  padding: 2em 3em;
+  padding: 2rem 3rem;
 }
 
 .link {
-  padding-right: 2em;
+  padding-right: 2rem;
 }
 
 .searchbar {
   width: 100%;
-  padding: 0 3%;
+  padding: 0 2.5rem;
 }
 
 input {
   width: 80%;
   height: 96%;
   background-color: #f0f0f0;
-  border-radius: 8px 0 0 8px;
+  border-radius: 0.5rem 0 0 0.5rem;
   border: none;
-  padding-left: 3%;
-  font-size: 0.9em;
+  padding-left: 2rem;
+  font-size: 0.9rem;
   font-weight: 200;
 }
 
 input::placeholder {
   font-family: 'Azeret Mono';
-  font-size: 0.9em;
+  font-size: 0.9rem;
   font-weight: 200;
 }
 
@@ -48,8 +48,8 @@ input:focus {
   width: 10%;
   height: 100%;
   background-color: #b3e0e2;
-  border-radius: 0 8px 8px 0;
+  border-radius: 0 0.5rem 0.5rem 0;
   border: none;
-  font-size: 1em;
+  font-size: 1rem;
   padding: 0;
 }
