+++ packages/frontend/src/components/infoCard.css
@@ -5,7 +5,7 @@
 }
 
 .card-heading {
-  padding: 35px 0 10px 0;
+  padding: 2rem 0 0.8rem 0;
 }
 
 .break {
