+++ packages/frontend/src/pages/AttesterPage.jsx
@@ -2,8 +2,9 @@ import React, { useContext, useState, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
 import { observer } from 'mobx-react-lite'
 import state from '../contexts/state'
-// import Tooltip from '../components/Tooltip'
+import Header from '../components/Header'
 import InfoCard from '../components/InfoCard'
+import EpochTime from '../components/EpochTime'
 import EpochView from '../components/EpochView'
 import UserView from '../components/UserView'
 import Footer from '../components/Footer'
@@ -35,7 +36,8 @@ export default observer(() => {
   const epochIds = [...(attester.epochsByAttesterId.get(attesterId) || [])]
   const lastEpoch = attester.epochsById.get(epochIds.pop())
   return (
-    <>
+    <div className="content">
+      <Header />
       <div className="container">
         <div className="left-container">
           <h3>Attester</h3>
@@ -89,7 +91,10 @@ export default observer(() => {
               heading="Total Bytes Given"
               value1={stats.totalBytes ?? 0}
             />
+            {deployment ? <EpochTime deployment={deployment} /> : null}
+            {deployment ? null : 'Loading...'}
           </div>
+
           {selectedView === 'Attestations' ? (
             <>
               <div style={{ display: 'flex' }}>
@@ -134,6 +139,6 @@ export default observer(() => {
       </div>
 
       <Footer />
-    </>
+    </div>
   )
 })
