+++ packages/frontend/src/pages/EpochKeyPage.jsx
@@ -1,8 +1,9 @@
 import React, { useContext, useEffect } from 'react'
-import { useParams, Link } from 'react-router-dom'
+import { useParams } from 'react-router-dom'
 import { observer } from 'mobx-react-lite'
 import state from '../contexts/state'
-import Tooltip from '../components/Tooltip'
+import Header from '../components/Header'
+import EpochKeyInfo from '../components/EpochKeyInfo'
 import EpochKeyEvent from '../components/EpochKeyEvent'
 import Footer from '../components/Footer'
 import dayjs from 'dayjs'
@@ -25,67 +26,30 @@ export default observer(() => {
   const attestations = unirep.attestationsByEpochKey.get(epochKeyId)
 
   return (
-    <>
+    <div className="content">
+      <Header />
       <div className="container">
         <div className="left-container">
           <h3>Epoch Key</h3>
-          <div className="info-card">
-            {attestations ? (
-              <>
-                <div className="flex">
-                  <h5>Attester</h5>
-                  <Link
-                    to={`/attester/0x${BigInt(
-                      attestations[0].attesterId
-                    ).toString(16)}`}
-                  >
-                    <h4>
-                      <span>
-                        {`0x${BigInt(attestations[0].attesterId).toString(
-                          16
-                        )}`.slice(0, 7)}
-                      </span>
-                      ...
-                      <span>
-                        {`0x${BigInt(attestations[0].attesterId).toString(
-                          16
-                        )}`.slice(-5)}
-                      </span>
-                    </h4>
-                  </Link>
-                </div>
-                <div className="flex">
-                  <h5>Epoch</h5>
-                  <h6>{attestations[0].epoch}</h6>
-                </div>
-                <div className="flex">
-                  <h5>Attestations Received</h5>
-                  <h6>{attestations.length}</h6>
-                </div>
-                <div className="flex">
-                  <h5>Epoch Key</h5>
-                  <Tooltip
-                    text="A user's temporary, random-value-like epochKey is the receiver of attestations."
-                    maxWidth={200}
-                  />
-                </div>
-                <h6 className="break">{id}</h6>
-              </>
-            ) : (
-              'Loading...'
-            )}
-          </div>
+          {attestations ? (
+            <EpochKeyInfo
+              attesterId={attestations[0].attesterId}
+              epoch={attestations[0].epoch}
+              numAttestations={attestations.length}
+              epochKey={id}
+            />
+          ) : null}
+          {attestations ? null : 'Loading...'}
         </div>
 
         <div className="right-container">
           <h3>Attestations</h3>
           <div className="flex events-header">
-            <h4>Epoch Status</h4>
             <h4>Field Index</h4>
             <h4>Change</h4>
             <h4>Timestamp</h4>
           </div>
-          <div className="scroll">
+          <div>
             {attestations
               ? attestations.map((attestation) => (
                   <EpochKeyEvent
@@ -98,8 +62,7 @@ export default observer(() => {
           </div>
         </div>
       </div>
-
       <Footer />
-    </>
+    </div>
   )
 })
